third_party/blink/web_tests/wpt_internal/css/css-size-adjust/text-size-adjust-none.html

<!DOCTYPE html>
<meta name="assert" content="text-size-adjust: none should disable text size adjustment.">
<link rel="help" href="https://drafts.csswg.org/css-size-adjust/#propdef-text-size-adjust">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<style>
  body {
    text-size-adjust: none;
  }
</style>

<body onload="runTest()">
  <div id="div">
    Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello
    Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello
    Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello Hello
    Hello Hello Hello Hello
  </div>
</body>

<script>
  function runTest() {

    test(() => {
      assert_equals(getComputedStyle(div)['font-size'], '16px');
    }, 'Text-size-adjust:none should disable adjustment');
  }
</script>
